+++ src/app/dishes/create/recipe-section-step.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { DishRecipeSectionWithId, DishRecipeStepWithId } from '@/src/types/dish.types';
-import { InputString } from '@/src/components/common/form/input-string';
+import { DebounceInputString } from '@/src/components/common/form/input-string';
 import { useState } from 'react';
 import { useRecipeFormContext } from '@/src/contexts/recipe-form.context';
 import { removeSectionStep, updateSectionStep } from '@/src/helpers/recipe-form.helper';
@@ -27,21 +27,17 @@ export function RecipeSectionStep({ section, step }: RecipeSectionStepProps) {
 
         const parsedStepResult = parseStep(newValue, translations);
 
-        // console.log(JSON.stringify(parsedStepResult));
-
-        // const stepText = 'Dodaj 100 g masła i 1 łyżeczka cukru, potem gotuj przez 5 minut. Dodaj cebulę i marchewkę.';
-        // const stepText = 'Dodaj marchew i pietruszkę, duś przez 10 minut, przypraw. Wlej szklankę cukru, gotuj jeszcze kwadrans. Dodaj jabłko, duś jeszcze 10 minut.';
-        // const result = parseFullStep(stepText, translations);
-
+        // TODO: The problem is that `cukru pudru` does not exist. The algorithm compares the actual text to its language occurrence...
         for (const stepText of [
-            'Dodaj 100 g masła i 2 łyżki cukru pudru do miski',
-            'Wymieszaj wszystko dokładnie przez 2 minuty.'
+            'Dodaj 100 g masło i 2 łyżki cukru pudru do miski',
+            'Wymieszaj wszystko dokładnie przez 2 minuty.',
+            'Dodaj marchew i pietruszkę, duś przez 10 minut, przypraw. Wlej szklankę cukru, gotuj jeszcze kwadrans. Dodaj jabłko, duś jeszcze 10 minut.'
         ]) {
+            console.time(stepText);
             const result = parseStep(stepText, translations);
             console.log(stepText, JSON.stringify(result));
+            console.timeEnd(stepText);
         }
-
-        // console.log(parseIngredients(stepText));
 
         setInstructionStep(newValue);
         onChangeSections(modifiedSections);
@@ -56,7 +52,14 @@ export function RecipeSectionStep({ section, step }: RecipeSectionStepProps) {
     return (
         <li className={styles['recipe-section-step']}>
             <div className={styles['step-container']}>
-                <InputString label={'Wpisz, co należy wykonać w tym kroku'} value={instructionStep} setValue={onChangeStep} width={'670px'} />
+                <DebounceInputString
+                    label={'Wpisz, co należy wykonać w tym kroku'}
+                    value={instructionStep}
+                    setValue={onChangeStep}
+                    width={'670px'}
+                    minLength={2}
+                    debounceTimeout={300}
+                />
                 <RemoveButton label={'Usuń'} onClick={onRemoveStep} />
             </div>
         </li>